coonection work

+++ PlantC.CitoyensEntreprise.DAL/Entities/Projet.cs
@@ -1,5 +1,7 @@
-﻿namespace PlantC.CitoyensEntreprise.DAL.Entities {
-    public class Projet {
+﻿namespace PlantC.CitoyensEntreprise.DAL.Entities
+{
+    public class Projet
+    {
 
         public int Id { get; set; }
         public string Reference { get; set; }
@@ -7,13 +9,13 @@
         public int? NbArbres { get; set; }
         public int? NbFruits { get; set; }
         public int? Metres { get; set; }
-        public double? Hectares { get; set; }
+        public decimal? Hectares { get; set; }
         public int IDLocalisation { get; set; }
-        public double TonnesCO2 { get; set; }
-        public double HeuresTravail { get; set; }
-        public double CoutDuProjet { get; set; }
+        public decimal TonnesCO2 { get; set; }
+        public decimal HeuresTravail { get; set; }
+        public decimal CoutDuProjet { get; set; }
 
-        public double Contribution { get; set; }
+        public decimal Contribution { get; set; }
         //Ne pas le stocker en DB, le calcul sera fait au fur et à mesure
 
     }
